requirements.txt
python-dotenv
google-genai
Pyaudio
fastapi
uvicorn
jinja2
